+++ src/views/Dashboard/Siswa/IndexView.vue
@@ -10,22 +10,7 @@
                         </div>
                     </div>
                     <div class="row">
-                        <div class="col-lg-3">
-                            <div class="card">                            
-                                <!-- <div class="card-body" style="position: relative;">
-                                    <div class="d-flex align-items-center justify-content-between">
-                                        <div>
-                                            <h3>$168.90</h3>
-                                            <span class="text-muted fw-semibold">This Month</span>
-                                        </div>
-                                        <div class="text-success fw-bold font-size-lg">+18%</div>
-                                        </div>
-                                    <div class="resize-triggers">
-                                        <div class="expand-trigger">
-                                        </div>
-                                    </div>
-                                </div> -->
-                            </div>
+                        <div class="col-lg-3">                           
                         </div>                        
                         <div class="col-lg-9">
                             <div class="">                            
@@ -85,6 +70,7 @@
 
 import axios from 'axios'
 import {onMounted, ref} from 'vue'
+import Swal from 'sweetalert2'
 
 export default {
     // 
@@ -101,16 +87,43 @@ export default {
             });
         });
 
-         function destroy(id,index){
-            axios.delete(
-                `/api/siswa/${id}`
-            )
-            .then(() => {
-                siswa.value.data.splice(index,1)
-                window.location.reload();
-            }).catch((err) => {
-                console.log(err.response.data);
-            });
+         async function destroy(id,index){
+             Swal.fire({
+                 title: 'Hapus Data Siswa',
+                 text: "Yakin dihapus ?",
+                 icon: 'warning',
+                 showCancelButton: true,
+                 confirmButtonColor: '#3085d6',
+                 cancelButtonColor: '#d33',
+                 confirmButtonText: 'Hapus !'
+             }).then((result) => {
+                 if (result.isConfirmed) {
+                     axios.delete(
+                         `/api/siswa/${id}`
+                     )
+                         .then(() => {
+                             siswa.value.data.splice(index, 1)
+                             Swal.fire({
+                                 title: 'Berhasil!',
+                                 text: 'Siswa Berhasil dihapus',
+                                 icon: 'success',
+                                 confirmButtonText: 'Lanjut !'
+                             }
+                             )
+                             window.location.reload();
+                         }).catch((err) => {
+                             console.log(err.response.data);
+                             Swal.fire({
+                                 title: 'Gagal!',
+                                 text: 'Siswa memiliki relasi dengan tabel lain !',
+                                 icon: 'error',
+                                 confirmButtonText: 'Lanjut !'
+                             }
+                             )
+                         });
+
+                 }
+             })           
         } 
 
         return {
